refactor(session): extract session graph export module (#186)
