+++ crates/pi-coding-agent/src/main.rs
@@ -11,6 +11,7 @@ mod provider_credentials;
 mod runtime_loop;
 mod session;
 mod session_commands;
+mod session_graph_commands;
 mod session_navigation_commands;
 mod skills;
 mod skills_commands;
@@ -102,12 +103,10 @@ pub(crate) use crate::provider_auth::{
     provider_api_key_candidates_with_inputs, provider_auth_capability, provider_auth_mode_flag,
     resolve_api_key,
 };
+#[cfg(test)]
+pub(crate) use crate::provider_credentials::resolve_store_backed_provider_credential;
 pub(crate) use crate::provider_credentials::{
     resolve_non_empty_secret_with_source, CliProviderCredentialResolver, ProviderCredentialResolver,
-};
-#[cfg(test)]
-pub(crate) use crate::provider_credentials::{
-    resolve_store_backed_provider_credential,
 };
 pub(crate) use crate::runtime_loop::{
     resolve_prompt_input, run_interactive, run_prompt, run_prompt_with_cancellation,
@@ -124,6 +123,12 @@ pub(crate) use crate::session_commands::{
     SESSION_SEARCH_DEFAULT_RESULTS, SESSION_SEARCH_PREVIEW_CHARS,
 };
 pub(crate) use crate::session_commands::{session_message_preview, session_message_role};
+pub(crate) use crate::session_graph_commands::execute_session_graph_export_command;
+#[cfg(test)]
+pub(crate) use crate::session_graph_commands::{
+    escape_graph_label, render_session_graph_dot, render_session_graph_mermaid,
+    resolve_session_graph_format, SessionGraphFormat,
+};
 #[cfg(test)]
 pub(crate) use crate::session_navigation_commands::{
     branch_alias_path_for_session, load_branch_aliases, load_session_bookmarks,
@@ -2416,6 +2421,42 @@ fn is_expired_unix(expires_unix: Option<u64>, now_unix: u64) -> bool {
     matches!(expires_unix, Some(value) if value <= now_unix)
 }
 
+pub(crate) fn write_text_atomic(path: &Path, content: &str) -> Result<()> {
+    if path.as_os_str().is_empty() {
+        bail!("destination path cannot be empty");
+    }
+    if path.exists() && path.is_dir() {
+        bail!("destination path '{}' is a directory", path.display());
+    }
+
+    let parent_dir = path
+        .parent()
+        .filter(|dir| !dir.as_os_str().is_empty())
+        .unwrap_or_else(|| Path::new("."));
+    std::fs::create_dir_all(parent_dir)
+        .with_context(|| format!("failed to create {}", parent_dir.display()))?;
+
+    let temp_name = format!(
+        ".{}.tmp-{}-{}",
+        path.file_name()
+            .and_then(|name| name.to_str())
+            .unwrap_or("session-graph"),
+        std::process::id(),
+        current_unix_timestamp()
+    );
+    let temp_path = parent_dir.join(temp_name);
+    std::fs::write(&temp_path, content)
+        .with_context(|| format!("failed to write temporary file {}", temp_path.display()))?;
+    std::fs::rename(&temp_path, path).with_context(|| {
+        format!(
+            "failed to rename temporary graph file {} to {}",
+            temp_path.display(),
+            path.display()
+        )
+    })?;
+    Ok(())
+}
+
 fn validate_session_file(cli: &Cli) -> Result<()> {
     if cli.no_session {
         bail!("--session-validate cannot be used together with --no-session");
@@ -2466,169 +2507,6 @@ fn initialize_session(agent: &mut Agent, cli: &Cli, system_prompt: &str) -> Resu
     }
 
     Ok(SessionRuntime { store, active_head })
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum SessionGraphFormat {
-    Mermaid,
-    Dot,
-}
-
-impl SessionGraphFormat {
-    fn as_str(self) -> &'static str {
-        match self {
-            Self::Mermaid => "mermaid",
-            Self::Dot => "dot",
-        }
-    }
-}
-
-fn resolve_session_graph_format(path: &Path) -> SessionGraphFormat {
-    let extension = path
-        .extension()
-        .and_then(|value| value.to_str())
-        .unwrap_or_default();
-    if extension.eq_ignore_ascii_case("dot") {
-        SessionGraphFormat::Dot
-    } else {
-        SessionGraphFormat::Mermaid
-    }
-}
-
-fn escape_graph_label(raw: &str) -> String {
-    raw.replace('\\', "\\\\").replace('"', "\\\"")
-}
-
-fn session_graph_node_label(entry: &crate::session::SessionEntry) -> String {
-    format!(
-        "{}: {} | {}",
-        entry.id,
-        session_message_role(&entry.message),
-        session_message_preview(&entry.message)
-    )
-}
-
-fn render_session_graph_mermaid(entries: &[crate::session::SessionEntry]) -> String {
-    let mut ordered = entries.iter().collect::<Vec<_>>();
-    ordered.sort_by_key(|entry| entry.id);
-
-    let mut lines = vec!["graph TD".to_string()];
-    if ordered.is_empty() {
-        lines.push("  empty[\"(empty session)\"]".to_string());
-        return lines.join("\n");
-    }
-
-    for entry in &ordered {
-        lines.push(format!(
-            "  n{}[\"{}\"]",
-            entry.id,
-            escape_graph_label(&session_graph_node_label(entry))
-        ));
-    }
-    for entry in &ordered {
-        if let Some(parent_id) = entry.parent_id {
-            lines.push(format!("  n{} --> n{}", parent_id, entry.id));
-        }
-    }
-    lines.join("\n")
-}
-
-fn render_session_graph_dot(entries: &[crate::session::SessionEntry]) -> String {
-    let mut ordered = entries.iter().collect::<Vec<_>>();
-    ordered.sort_by_key(|entry| entry.id);
-
-    let mut lines = vec!["digraph session {".to_string(), "  rankdir=LR;".to_string()];
-    if ordered.is_empty() {
-        lines.push("  empty [label=\"(empty session)\"];".to_string());
-    } else {
-        for entry in &ordered {
-            lines.push(format!(
-                "  n{} [label=\"{}\"];",
-                entry.id,
-                escape_graph_label(&session_graph_node_label(entry))
-            ));
-        }
-        for entry in &ordered {
-            if let Some(parent_id) = entry.parent_id {
-                lines.push(format!("  n{} -> n{};", parent_id, entry.id));
-            }
-        }
-    }
-    lines.push("}".to_string());
-    lines.join("\n")
-}
-
-fn render_session_graph(
-    format: SessionGraphFormat,
-    entries: &[crate::session::SessionEntry],
-) -> String {
-    match format {
-        SessionGraphFormat::Mermaid => render_session_graph_mermaid(entries),
-        SessionGraphFormat::Dot => render_session_graph_dot(entries),
-    }
-}
-
-fn write_text_atomic(path: &Path, content: &str) -> Result<()> {
-    if path.as_os_str().is_empty() {
-        bail!("destination path cannot be empty");
-    }
-    if path.exists() && path.is_dir() {
-        bail!("destination path '{}' is a directory", path.display());
-    }
-
-    let parent_dir = path
-        .parent()
-        .filter(|dir| !dir.as_os_str().is_empty())
-        .unwrap_or_else(|| Path::new("."));
-    std::fs::create_dir_all(parent_dir)
-        .with_context(|| format!("failed to create {}", parent_dir.display()))?;
-
-    let temp_name = format!(
-        ".{}.tmp-{}-{}",
-        path.file_name()
-            .and_then(|name| name.to_str())
-            .unwrap_or("session-graph"),
-        std::process::id(),
-        current_unix_timestamp()
-    );
-    let temp_path = parent_dir.join(temp_name);
-    std::fs::write(&temp_path, content)
-        .with_context(|| format!("failed to write temporary file {}", temp_path.display()))?;
-    std::fs::rename(&temp_path, path).with_context(|| {
-        format!(
-            "failed to rename temporary graph file {} to {}",
-            temp_path.display(),
-            path.display()
-        )
-    })?;
-    Ok(())
-}
-
-fn execute_session_graph_export_command(runtime: &SessionRuntime, command_args: &str) -> String {
-    let destination = PathBuf::from(command_args.trim());
-    let format = resolve_session_graph_format(&destination);
-    let graph = render_session_graph(format, runtime.store.entries());
-    let nodes = runtime.store.entries().len();
-    let edges = runtime
-        .store
-        .entries()
-        .iter()
-        .filter(|entry| entry.parent_id.is_some())
-        .count();
-
-    match write_text_atomic(&destination, &graph) {
-        Ok(()) => format!(
-            "session graph export: path={} format={} nodes={} edges={}",
-            destination.display(),
-            format.as_str(),
-            nodes,
-            edges
-        ),
-        Err(error) => format!(
-            "session graph export error: path={} error={error}",
-            destination.display()
-        ),
-    }
 }
 
 fn format_id_list(ids: &[u64]) -> String {
